refactored location service
